Separated converters to operations. Added Relu support
