+++ pt2keras/core/onnx/convert/ops.py
@@ -7,31 +7,6 @@ from .common import converter
 from ..util import tensor_proto_to_tf_dtype
 
 
-@converter('Conv')
-def add(node: onnx.NodeProto, input_layer, *node_inputs):
-    """
-    Convert the add operation
-    Args:
-        node: The node that we wish to convert
-    Returns:
-
-    """
-    attributes: t.Dict = node.attributes
-    weights_shape = node.weights[0].shape
-    filter_count = weights_shape[-1]
-    outputs = keras.layers.Conv2D(
-        filter_count,  # filters
-        attributes['kernel_shape'],  # Kernel size
-        strides=attributes['strides'],
-        groups=attributes['group'],
-        weights=node.weights,
-        dilation_rate=attributes['dilations'],
-        # Weights is of length two ['weights', 'bias']
-        use_bias=len(node.weights) == 2,
-    )(input_layer)
-    return outputs
-
-
 @converter('Constant')
 def constant(node: onnx.NodeProto, input_layer, *inputs):
     """
@@ -61,14 +36,8 @@ def divide(node: onnx.NodeProto, input_layer, lhs, rhs):
 @converter('Cast')
 def cast(node: onnx.NodeProto, input_layer, *args):
     """
-    Floor divide is considered a Cast operation in onnx
-    Args:
-        node:
-        computational_graph:
-        current_inputs:
-
-    Returns:
-
+    Floor divide is considered a Cast operation in onnx,
+    since we are casting from float32 to int
     """
     tf_dtype = tensor_proto_to_tf_dtype(node.attributes['to'])
     outputs = K.cast(input_layer, dtype=tf_dtype)
